fix: selectively merge audit fixes without wire-breaking or token logging
Take the high-value concurrency/security fixes from fix/audit-critical-issues:
config/state_sync lock discipline, distributed lock atomic ops, metrics RMW,
epoll read accounting, mysql identifier sanitize, room/player locks, login
token validator hook, http body limit, rpc buffer races, timer/frame snapshot
eviction, and related medium fixes.

Deliberately excluded:
- LE->BE int64/uint64 timestamp change (wire-breaking for existing clients)
- token value in login INFO log
- file mode 100755 noise
- incomplete/cosmetic-only test mode churn

Also fix FrameSyncComponent self-deadlock: broadcast outside the mutex so
broadcast_frame_state's lock is not re-entered from timer/input paths.

+++ include/chwell/net/epoll_demuxer.h
@@ -130,7 +130,7 @@ public:
         }
 
         struct epoll_event ev{};
-        ev.events = static_cast<uint32_t>(events) | (edge_trigger ? EPOLLET : 0);
+        ev.events = static_cast<uint32_t>(events) | (edge_trigger ? static_cast<uint32_t>(EPOLLET) : 0u);
         ev.data.fd = fd;
         return ::epoll_ctl(epoll_fd_, EPOLL_CTL_ADD, fd, &ev) >= 0;
     }
@@ -150,7 +150,7 @@ public:
         }
 
         struct epoll_event ev{};
-        ev.events = static_cast<uint32_t>(events) | (et ? EPOLLET : 0);
+        ev.events = static_cast<uint32_t>(events) | (et ? static_cast<uint32_t>(EPOLLET) : 0u);
         ev.data.fd = fd;
         return ::epoll_ctl(epoll_fd_, EPOLL_CTL_MOD, fd, &ev) >= 0;
     }
